requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0
tqdm>=4.64.0

# XML processing
lxml>=4.9.0
beautifulsoup4>=4.11.0

# Embeddings and ML
sentence-transformers>=2.2.0
transformers>=4.20.0
torch>=2.0.0

# Vector database
faiss-cpu>=1.7.4

# LLM integration
requests>=2.28.0
openai>=1.0.0  # For OpenAI-compatible APIs

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Development and testing
jupyter>=1.0.0
ipykernel>=6.15.0
pytest>=7.1.0

# Optional: For advanced text processing
spacy>=3.4.0
nltk>=3.8